// src/types/apple-ads.ts
export interface Money {
  amount: string;
  currency: string;
}

export interface Campaign {
  id: number;
  orgId: number;
  name: string;
  budgetAmount: Money | null;
  dailyBudgetAmount: Money | null;
  adamId: number;
  countriesOrRegions: string[] | null;
  status: "ENABLED" | "PAUSED";
  servingStatus: "RUNNING" | "NOT_RUNNING";
  displayStatus: "RUNNING" | "ON_HOLD" | "PAUSED" | "DELETED";
  supplySources: string[] | null;
  adChannelType: string;
  billingEvent: string;
  paymentModel?: "LOC" | "PAYG";
  deleted: boolean;
  startTime: string;
  endTime?: string | null;
  modificationTime: string;
}
// ...
export interface Ad {
  id: number;
  orgId: number;
  campaignId: number;
  adGroupId: number;
  name: string;
  creativeId: number;
  status: "ENABLED" | "PAUSED";
  servingStatus: string;
  creationTime: string;
  modificationTime: string;
  deleted: boolean;
  rejectionReasons?: Array<{ reason: string; message: string }>;
}

export interface BudgetOrder {
  id: number;
  name: string;
  budget: Money | null;
  status: "ACTIVE" | "INACTIVE" | "EXHAUSTED" | "CANCELLED";
  startDate: string;
  endDate?: string | null;
  orderNumber?: string | null;
  supplySources?: string[] | null;
  modificationTime: string;
}

export interface Creative {
  id: number;
  orgId: number;
  adamId: number;
  name: string;
  type: string;
  state: string;
  creationTime: string;
  modificationTime: string;
}